deploy multiple updates

+++ client/src/pages/admin-view/products.jsx
@@ -1,5 +1,5 @@
 import ProductImageUpload from "@/components/admin-view/image-upload";
-import AdminProductTile from "@/components/admin-view/product-title";
+import AdminProductTile from "@/components/admin-view/product-tile";
 // import AdminProductTile from "@/components/admin-view/product-tile";
 import CommonForm from "@/components/common/form";
 import { Button } from "@/components/ui/button";
@@ -11,15 +11,18 @@ import {
 } from "@/components/ui/sheet";
 import { addProductFormElements } from "@/config";
 import { useToast } from "@/hooks/use-toast";
+import { fetchAllCategories } from "@/store/admin/categories-slice"; //alias as adminCategoriesSlice to be used in store.js
+import { fetchAllBrands } from "@/store/admin/brands-slice";
 // import { useToast } from "@/components/ui/use-toast";
 import {
   addNewProduct,
   deleteProduct,
   editProduct,
   fetchAllProducts,
-} from "@/store/admin/products-slice";
+} from "@/store/admin/products-slice"; //alias as AdminProductsSlice to be used in store.js
 import { Fragment, useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
+import Swal from "sweetalert2";
 
 const initialFormData = {
   image: null,
@@ -42,7 +45,7 @@ function AdminProducts() {
   const [imageLoadingState, setImageLoadingState] = useState(false);
   const [currentEditedId, setCurrentEditedId] = useState(null); //store current id for item edited
 
-  const { productList } = useSelector((state) => state.adminProducts);
+  const { productList } = useSelector((state) => state.adminProducts); // adminProducts value from store.js
   const dispatch = useDispatch();
   const { toast } = useToast();
 
@@ -78,17 +81,38 @@ function AdminProducts() {
             setImageFile(null); // reset the image file
             setFormData(initialFormData); //reset the form data
             toast({
-              title: "Product add successfully",
+              title: "Product added successfully",
             });
           }
         });
   }
 
   function handleDelete(getCurrentProductId) {
-   // console.log(getCurrentProductId);
-    dispatch(deleteProduct(getCurrentProductId)).then((data) => {
-      if (data?.payload?.success) {
-        dispatch(fetchAllProducts());   //get list of the products
+    //add here the swal sweet alert dialog
+    Swal.fire({
+      title: `Are you sure? delete Item : ${getCurrentProductId}`,
+      // text: "You won't be able to revert this!",
+      // icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Delete Now!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        // console.log(getCurrentProductId);
+        dispatch(deleteProduct(getCurrentProductId)).then((data) => {
+          if (data?.payload?.success) {
+            dispatch(fetchAllProducts()); //get list of the products
+            toast({
+              title: "Product deleted successfully",
+            });
+          }
+        });
+        // Swal.fire({
+        //   title: "Deleted!",
+        //   text: "Product has been deleted.",
+        //   icon: "success",
+        // });
       }
     });
   }
@@ -103,9 +127,20 @@ function AdminProducts() {
   useEffect(() => {
     dispatch(fetchAllProducts());
   }, [dispatch]);
-
   console.log(formData, "formData");
   //console.log(productList, uploadedImageUrl, "productList");
+
+  const { categoryList } = useSelector((state) => state.adminCategories);
+  useEffect(() => {
+    dispatch(fetchAllCategories());
+  }, [dispatch]);
+  console.log(categoryList, "categoryList for select");
+
+  const { brandList } = useSelector((state) => state.adminBrands);
+  useEffect(() => {
+    dispatch(fetchAllBrands());
+  }, [dispatch]);
+  console.log(brandList, "brandList for select");
 
   return (
     <Fragment>
@@ -155,12 +190,13 @@ function AdminProducts() {
           />
 
           <div className="py-6">
+            {/* reusable component */}
             <CommonForm
               onSubmit={onSubmit}
               formData={formData}
               setFormData={setFormData}
               buttonText={currentEditedId !== null ? "Edit" : "Add"}
-              formControls={addProductFormElements}
+              formControls={addProductFormElements(categoryList,brandList)}
               isBtnDisabled={!isFormValid()}
             />
           </div>
